returning records for given date
What changes you have made?
in query added date as valid argument and gives back records for given
empId

Why this change is necessary?
this was necessary for implementing date feature for query

What is the impact of the change?
now my app can return a records of the particular date